auth_service/requirements.txt
Django==4.2.2
django-redis==5.3.0
djangorestframework==3.14.0
PyJWT==2.7.0
django-cors-headers==4.1.0
psycopg2-binary==2.9.6